backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch>=2.2.0
transformers>=4.40.0
diffusers>=0.30.0
accelerate>=0.20.0
Pillow==10.0.1
requests==2.31.0
numpy<2.0
